+++ pkg/chatbot/playback.go
@@ -1,6 +1,7 @@
 package chatbot
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"strconv"
@@ -13,7 +14,6 @@ import (
 	"github.com/adanalife/tripbot/pkg/helpers"
 	"github.com/adanalife/tripbot/pkg/users"
 	"github.com/adanalife/tripbot/pkg/video"
-	vlcClient "github.com/adanalife/tripbot/pkg/vlc-client"
 )
 
 // lastTimewarpTime is used to rate-limit users so they can't
@@ -27,7 +27,7 @@ func (a *App) timewarp() {
 	a.Onscreens.ShowTimewarp()
 
 	// shuffle to a new video
-	err := vlcClient.PlayRandom()
+	err := a.VLC.PlayRandom()
 	if err != nil {
 		terrors.Log(err, "error from VLC client")
 	}
@@ -37,7 +37,7 @@ func (a *App) timewarp() {
 	lastTimewarpTime = time.Now()
 }
 
-func (a *App) timewarpCmd(user *users.User, _ []string) {
+func (a *App) timewarpCmd(ctx context.Context, user *users.User, _ []string) {
 	log.Println(user.Username, "ran !timewarp")
 
 	// exit early if we're on OS X
@@ -63,7 +63,7 @@ func (a *App) timewarpCmd(user *users.User, _ []string) {
 	a.timewarp()
 }
 
-func (a *App) jumpCmd(user *users.User, params []string) {
+func (a *App) jumpCmd(ctx context.Context, user *users.User, params []string) {
 	var err error
 	log.Println(user.Username, "ran !jump")
 
@@ -106,7 +106,7 @@ func (a *App) jumpCmd(user *users.User, params []string) {
 		return
 	}
 	// tell VLC to play it
-	err = vlcClient.PlayFileInPlaylist(randomVid.File())
+	err = a.VLC.PlayFileInPlaylist(randomVid.File())
 	if err != nil {
 		terrors.Log(err, "error from VLC client")
 		Say("Usage: !jump [state]")
@@ -121,7 +121,7 @@ func (a *App) jumpCmd(user *users.User, params []string) {
 	lastTimewarpTime = time.Now()
 }
 
-func (a *App) skipCmd(user *users.User, params []string) {
+func (a *App) skipCmd(ctx context.Context, user *users.User, params []string) {
 	var err error
 	var n int
 	log.Println(user.Username, "ran !skip")
@@ -156,7 +156,7 @@ func (a *App) skipCmd(user *users.User, params []string) {
 	}
 
 	// skip to a new video
-	err = vlcClient.Skip(n)
+	err = a.VLC.Skip(n)
 	if err != nil {
 		terrors.Log(err, "error from VLC client")
 	}
@@ -166,7 +166,7 @@ func (a *App) skipCmd(user *users.User, params []string) {
 	lastTimewarpTime = time.Now()
 }
 
-func (a *App) backCmd(user *users.User, params []string) {
+func (a *App) backCmd(ctx context.Context, user *users.User, params []string) {
 	var err error
 	var n int
 	log.Println(user.Username, "ran !back")
@@ -201,7 +201,7 @@ func (a *App) backCmd(user *users.User, params []string) {
 	}
 
 	// back to an old video
-	err = vlcClient.Back(n)
+	err = a.VLC.Back(n)
 	if err != nil {
 		terrors.Log(err, "error from VLC client")
 	}